+++ src/propedit/propedit.h
@@ -19,3 +19,4 @@
 #define FILE_NOT_MODIFIED Y("The file has not been modified.")
 
 extern std::unique_ptr<mtx::doc_type_version_handler_c> g_doc_type_version_handler;
+extern std::unordered_map<uint64_t, uint64_t> g_track_uid_changes;